Refactor streaming interface and add Message type
