--- sdks/go/pkg/st2138/stream.go
@@ -1,61 +1,0 @@
-/*
- * Copyright 2026 Ross Video Ltd
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its
- * contributors may be used to endorse or promote products derived from this
- * software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-
-/**
- * @brief Streaming primitives for the Catena SDK.
- * @file stream.go
- * @copyright Copyright © 2026 Ross Video Ltd
- */
-
-package st2138
-
-import (
-	"google.golang.org/protobuf/proto"
-)
-
-// Message is implemented by the wrapper types that can be sent as a single
-// chunk of a streamed response. Wire returns the fully-formed protobuf message
-// for that chunk - the transport-neutral wire representation.
-//
-// A Message is self-contained: it carries everything the transport needs to
-// emit one chunk. Transports never reach past Wire to inspect wrapper internals
-// - the gRPC path sends the returned proto directly, and the REST path applies
-// its own JSON marshalling rules to it.
-type Message interface {
-	Wire() proto.Message
-}
-
-// Stream is the handler-facing sink for a streamed response. Each endpoint's
-// stream is parameterized by its own chunk type, so a handler can only send the
-// chunk type that endpoint expects. Send delivers one chunk to the client and
-// returns an error if the chunk could not be sent.
-type Stream[T Message] interface {
-	Send(chunk T) error
-}
